drm/dp/mst: handle connection status notify better.
If we get a CSN message from a DP1.2 downstream device, the
mst layer would attempt to send msgs to the device from the
irq, so could never receive replies. Set a flag on the port
and use the work queue to update values the port requires.

MST spec is crappy here, initial port info has guid/dpcd_rev
but CSN doesn't have it, so we manually probe it here.

Signed-off-by: Dave Airlie <airlied@redhat.com>