drm/amd/dal: Fix issue where unused HW is not powered down after boot
[Description]
init_hw is called on dc create and on set power state call,
but what appears incorrect is that it is setting accelerated
mode bit in scratch register.

We actually expect accelerated mode bit to be set on first commit,
when we actually enter accelerated mode. Entering accelerated
mode actually involves a handoff between VBIOS controlled
programming of display pipe to Driver control. During this process,
we want to init and power down all existing HW.

Power down of HW in enter accelerated mode was also not very
correct, since it currently takes the first commit context. But what
we really wanted to do is power everything off.

Finally, we need to do power gate of pipe when unused, but also
ungate when enabling the pipe.

Signed-off-by: Harry Wentland <harry.wentland@amd.com>
Acked-by: Harry Wentland <harry.wentland@amd.com>