[RFC] drm/amdgpu: drop DVO encoders support.
I'm a bit confused whether this is needed so it would be good to
confirm if hw shipped with it.

DAL as currently is doesn't have any support for this, it doesn't
have any encoder ids for DVO that I can see.

So before porting to DAL graphics objects, I'm removing this code,
so we can see if any regressions land here.

Signed-off-by: Dave Airlie <airlied@redhat.com>