drm: remove connector from lists in unregister.
So at the moment we remove the connector from the lists in the
destroy, but this means due to a lack of locking we can have
zombie connectors sitting on the list with kref 0.

Now we suck at locking connector lists in most drivers, but I
think for drivers that expect to use MST we should start being
pretty strict on locking for the connector list and then rely
on this to remove any unneeded connectors.

The main reason for moving this earlier is we know we can take
the locks here as the MST code for removing connectors runs
in a separate work queue, so we won't have any other locks held
going into the unregister, whereas the connector destroy on unref
could be happening from any context where we take a connector
reference.

Signed-off-by: Dave Airlie <airlied@redhat.com>